fix(react-service): 修复 build 命令构建之后路径错误的 bug
