+++ packages/serendipity-service-react/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@babel/core": "^7.12.10",
+    "@babel/preset-env": "^7.12.13",
     "@babel/preset-react": "^7.12.10",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.4.3",
     "@types/webpack-dev-server": "^3.11.1",
